moved sync method in Policy class
