+++ streamflow/recovery/rollback_recovery.py
@@ -252,122 +252,6 @@ class RollbackDeterministicWorkflowPolicy:
                     port_names.add(port_name)
         return list(port_names)
 
-    # todo: move it in recovery.py
-    async def sync_running_jobs(self, workflow):
-        logger.debug(f"INIZIO sync (wf {workflow.name}) RUNNING JOBS")
-        map_job_port = {}
-        job_token_names = [
-            token.value.name
-            for token in self.token_instances.values()
-            if isinstance(token, JobToken)
-        ]
-        logger.debug(f"TOKEN_GRAPH: {self.dag_tokens}")
-        # logger.debug(f"PORT_GRAPH: {self.dcg_port}")
-        str_token_tags = "\n".join(
-            [
-                f"{self.token_instances[k].tag if k in self.token_instances.keys() else k}: {[self.token_instances[v].tag if v in self.token_instances.keys() else v for v in values]}"
-                for k, values in self.dag_tokens.items()
-            ]
-        )
-        logger.debug(f"TOKEN_TAGS: {str_token_tags}")
-        logger.debug(
-            f"PORT_TAGS: { { k : [self.token_instances[v].tag for v in values] for k, values in self.port_tokens.items()} }"
-        )
-
-        for job_token in [
-            token
-            for token in self.token_instances.values()
-            if isinstance(token, JobToken)
-        ]:
-            job_request = await self.context.failure_manager.setup_job_request(
-                job_token.value.name, default_is_running=False
-            )
-            async with job_request.lock:
-                if job_request.is_running:
-                    logger.debug(
-                        f"Sync Job {job_token.value.name} (wf {workflow.name}): JobRequest is running on wf {job_request.workflow.name}. Other jobs: {job_token_names}"
-                    )
-                    output_port_names = await self.get_execute_output_port_names(
-                        job_token
-                    )
-                    logger.debug(f"Lista output_port_names: {output_port_names}")
-                    for output_port_name in output_port_names:
-                        port_recovery = self.context.failure_manager.add_waiter(
-                            job_token.value.name,
-                            output_port_name,
-                            workflow,
-                            map_job_port.get(job_token.value.name, None),
-                        )
-                        logger.debug(
-                            f"Created port {port_recovery.port.name} for wf {workflow.name} with waiting token {port_recovery.waiting_token}"
-                        )
-                        map_job_port.setdefault(job_token.value.name, port_recovery)
-                        workflow.add_port(port_recovery.port)
-                    if len(job_token_names) == 1:
-                        logger.debug(f"New-workflow {workflow.name} will be empty.")
-                        pass
-                        # raise FailureHandlingException(
-                        #     "There is only job job in this rollback, but it is already running in another workflow. Something is wrong."
-                        # )
-                    execute_step_out_token_ids = await get_execute_step_out_token_ids(
-                        self.dag_tokens.succ(job_token.persistent_id),
-                        self.context,
-                    )
-                    logger.debug(
-                        f"Lista execute_step_out_token_ids: {execute_step_out_token_ids}"
-                    )
-                    for t_id in execute_step_out_token_ids:
-                        for t_id_dead_path in self.remove_token_prev_links(t_id):
-                            self.remove_token(t_id_dead_path)
-
-                elif job_request.token_output and all(
-                    [
-                        await _is_token_available(t, self.context)
-                        for t in job_request.token_output.values()
-                    ]
-                ):
-                    # search execute token after job token, replace this token with job_requ token. then remove all the prev tokens
-                    logger.debug(
-                        f"Sync Job {job_token.value.name} (wf {workflow.name}): JobRequest has token_output { { k : v.persistent_id for k, v in job_request.token_output.items()} }"
-                    )
-                    for port_name in await self.get_execute_output_port_names(
-                        job_token
-                    ):
-                        new_token = job_request.token_output[port_name]
-                        port_row = await self.context.database.get_port_from_token(
-                            new_token.persistent_id
-                        )
-                        step_rows = await get_steps_from_output_port(
-                            port_row["id"], self.context
-                        )
-                        logger.debug(
-                            f"Sync Job {job_token.value.name} (wf {workflow.name}): Replace {self.get_equal_token(port_name, new_token)} with {new_token.persistent_id}"
-                        )
-                        await self.replace_token_and_remove(
-                            port_name,
-                            ProvenanceToken(
-                                new_token,
-                                TokenAvailability.Available,
-                                port_row,
-                                step_rows,
-                            ),
-                        )
-                else:
-                    logger.debug(
-                        f"Sync Job {job_token.value.name} (wf {workflow.name}): JobRequest set to running and job_token and token_output to None."
-                        f"\n\t- Prev value job_token: {job_request.job_token}\n\t- Prev value token_output: {job_request.token_output}"
-                    )
-                    job_request.is_running = True
-                    job_request.job_token = None
-                    job_request.token_output = None
-                    job_request.workflow = workflow
-                    await self.context.failure_manager.update_job_status(
-                        job_token.value.name, job_request.lock
-                    )
-            # end lock
-        # end for job token
-        logger.debug(f"FINE sync (wf {workflow.name}) RUNNING JOBS")
-
     # rename it in get_equal_token_id ?
     def get_equal_token(self, port_name, token):
         for t_id in self.port_tokens.get(port_name, []):
